+++ src/I18n/I18n.php
@@ -8,11 +8,11 @@ class I18n
 {
 	public function register(Loader $loader): void
 	{
-		$loader->action('plugins_loaded', $this, 'load_textdomain');
+		$loader->action('plugins_loaded', [$this, 'load_textdomain']);
 	}
 
 	public function load_textdomain(): void
 	{
-		load_plugin_textdomain('wp-plugin-boilerplate', false, dirname(plugin_basename(__FILE__), 2) . '/languages');
+		\load_plugin_textdomain('wp-plugin-boilerplate', false, dirname(plugin_basename(__FILE__), 2) . '/languages');
 	}
 }